@@ -7,7 +7,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>3.1.2</version>
+		<version>3.2.1</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 
@@ -18,11 +18,11 @@
 	<description>API de delivery de comida</description>
 
 	<properties>
-		<java.version>17</java.version>
-		<modelmapper.version>3.1.1</modelmapper.version>
-		<jasperreports.version>6.20.5</jasperreports.version>
-		<aws-java-sdk-s3.version>1.12.533</aws-java-sdk-s3.version>
-		<springdoc.version>2.2.0</springdoc.version>
+		<java.version>21</java.version>
+		<modelmapper.version>3.2.0</modelmapper.version>
+		<jasperreports.version>6.21.0</jasperreports.version>
+		<aws-java-sdk-s3.version>1.12.623</aws-java-sdk-s3.version>
+		<springdoc.version>2.3.0</springdoc.version>
 		<logback-ext-loggly.version>0.1.5</logback-ext-loggly.version>
 	</properties>
 
