@@ -13,7 +13,7 @@ public class ApiRetirementHandler implements HandlerInterceptor {
 	@Override
 	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
 			throws Exception {
-		if (request.getRequestURI().startsWith("/v1/cidades")) {
+		if (request.getRequestURI().startsWith("/v2/cidades")) {
 			response.setStatus(HttpStatus.GONE.value());
 			return false;
 		}
